backend/pyproject.toml
[project]
name = "recruit-backend"
version = "0.1.0"
description = "Take-home assignment backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.6",
    "redis>=5.2.0",
    "python-multipart>=0.0.20",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.1",
    "pyright>=1.1.396",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pyright]
typeCheckingMode = "standard"

